Linux Scripts update for VMs

+++ windows/cloud.ps1
@@ -155,8 +155,10 @@ if ($productivity_install) {
 }
 
 # Clean PowerShell install script and restart
-Write-Output "Cleaning up Utils script script"
-Remove-Item -Path $utils_script_name -Confirm:$false -ErrorAction SilentlyContinue
+If ($env:LOCAL  -eq 'no')  {
+    Write-Output "Cleaning up Utils script script"
+    Remove-Item -Path $utils_script_name -Confirm:$false -ErrorAction SilentlyContinue
+}
 
 if ($LOCAL -eq 'no')  {
     Restart-Computer -Force
